remove validateupstream package and move files to xds package
